+++ packages/docs/site/src/routes/components/_examples/ui/index.js
@@ -1,7 +1,9 @@
 import * as AlphabetPicker from './AlphabetPicker.js';
+import * as CenteredView from './CenteredView.js';
 import * as CopyToClipboard from './CopyToClipboard.js';
 import * as HighlightedText from './HighlightedText.js';
 import * as Icon from './Icon.js';
+import * as MessageView from './MessageView.js';
 import * as Link from './Link.js';
 import * as LinkButton from './LinkButton.js';
 import * as LoadingView from './LoadingView.js';
@@ -11,8 +13,10 @@ import {examplesFormatter3} from '../utils.js';
 
 export default examplesFormatter3([
 	AlphabetPicker,
+	CenteredView,
 	CopyToClipboard,
 	HighlightedText,
+	MessageView,
 	Icon,
 	Link,
 	LinkButton,